win32 fixes
Originally committed as revision 78 to svn://svn.ffmpeg.org/ffmpeg/trunk

+++ libav/avio.h
@@ -1,6 +1,6 @@
 /* output byte stream handling */
 
-typedef long long offset_t;
+typedef INT64 offset_t;
 
 /* unbuffered I/O */
 
@@ -87,8 +87,8 @@ int init_put_byte(ByteIOContext *s,
 
 void put_byte(ByteIOContext *s, int b);
 void put_buffer(ByteIOContext *s, unsigned char *buf, int size);
-void put_le64(ByteIOContext *s, unsigned long long val);
-void put_be64(ByteIOContext *s, unsigned long long val);
+void put_le64(ByteIOContext *s, UINT64 val);
+void put_be64(ByteIOContext *s, UINT64 val);
 void put_le32(ByteIOContext *s, unsigned int val);
 void put_be32(ByteIOContext *s, unsigned int val);
 void put_le16(ByteIOContext *s, unsigned int val);
@@ -105,12 +105,12 @@ void put_flush_packet(ByteIOContext *s);
 int get_buffer(ByteIOContext *s, unsigned char *buf, int size);
 int get_byte(ByteIOContext *s);
 unsigned int get_le32(ByteIOContext *s);
-unsigned long long get_le64(ByteIOContext *s);
+UINT64 get_le64(ByteIOContext *s);
 unsigned int get_le16(ByteIOContext *s);
 
 unsigned int get_be16(ByteIOContext *s);
 unsigned int get_be32(ByteIOContext *s);
-unsigned long long get_be64(ByteIOContext *s);
+UINT64 get_be64(ByteIOContext *s);
 
 extern inline int url_is_streamed(ByteIOContext *s)
 {
